cmake: update compilation options
This patch updated the compilation options used making them more close to
what is used when using the legacy Makefile. Some of them are really needed
when using mingw to compile the project which means this is also a
preparation patch to support mingw builds with cmake.

On top of this, 'add_compile_options()' is used for compiler options and
'add_definitions()' is used for preprocessor stuff as it should be. Since
it's a very tiny change, the check for GNU_GCC is also removed since at
this point no other compiler was really tested.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>