cmake: include directories on a per target basis
This patch makes use of the more recent 'target_include_directories()'
cmake api to only include the directories needed for eact target. This is
not entirely true for plugins as all plugins are seen as one big target
which means that we include all the directories needed to compile all
plugins.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>