cmake: remove git information from {lib}osc targets
There's no added value in appending git information to the version. This
information can already be seen in the app "About" page. For libosc, let's
just use the version defined in cmake (even though it's not that
meaningful). For osc let's remove it altogether as it would be a problem
when compiling with mingw. For dlls, cmake handles things correctly and
does not create any symlink. For apps it does creates them which won't work
on Windows.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>