cmake: add support for mingw/Windows builds
This patch leaves cmake prepared to be used by our CI scripts so that we
can cross compile osc for Windows (assuming all dependencies are present
and also cross compiled). This patch also makes it consistent in
identifying that we are building for windows. As stated in cmake docs, the
best way to do this is to look at 'CMAKE_SYSTEM_NAME'.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>