ci: appveyor: adapt build script to use cmake
This updates the script to use cmake. All the unused stuff was also
removed.

For the debian specific stuff, I'm not sure we still need to support it.
Anyways, I just removed it since if we do need to support it we need to
support it in cmake (bringing back the 'link_directories()' usage for
example) and not in the script.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>