cmake: use profiles basename
For the autogenerated profiles, the full path was being used when
'configure_file()' was called. This was causing cmake to create this full
directory inside profiles/. Hence, we need to extract the profile basename
and use this. In the end, things were still working properly so this is
just an improvement rather than a real fix.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>