plugins: adrv9002: fix segfault with dac buffering
The DAC Manager holds a weak reference to the IIO context (passed by the
plugin) which it uses to handle the dac devices (eg: creating buffers). In
dac_data_manager_free(), the Manager will destroy any buffers that were
created which means we cannot destroy the IIO context before calling
this function...

While at it, reordered how resources are freed during the plugin
destructor so that we free things in the reverse order they were created
(as it should be typically done).

Signed-off-by: Nuno Sá <nuno.sa@analog.com>