plugins: adrv9002: update interface gain combo box
With [1], the interface gain available attribute can change on profile
load. Hence, we also need to reinitialize the combo box to reflect that.

Special care is needed to not get a 'changed' signal on the widget when
clearing the combo box. Basically we block the signal before clearing
the box.

[1]: https://github.com/analogdevicesinc/linux/pull/2126
Signed-off-by: Nuno Sá <nuno.sa@analog.com>