plugins: adrv9002: move widgets to signal blocked API
With this change, we should not get into any registered signal handler
without an explicit user interaction. As a result, custom handlers as
the RX gain are not needed anymore. We had it, since moving the gain
control to automatic could lead to a gain change that would, in turn, lead
to the signal handler which would try to save the same value. Moreover,
as the gain control mode is automatic, the driver would return an error.
Now, updating the gain (in the update_display() timer) is done with signals
blocked which means there's no need to worry about undesired behavior.

Note the combo boxes were not changed. That will be done in a follow-up
patch together with dropping the custom combo box implementation.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>