github/workflow: creates deb for Ubuntu 22.04
Include files to generate debs, targeting Ubuntu 22.04.
Uses the ENABLE_PACKAGING cmake flag to generate the packages.
Uses dependencies available in the main package manager repositories.
The workflow runs in master, this branch and on published release.
Can also create rpm, but this is not uploaded by the workflow.
Note: Ubuntu libcurl4 flavor is openssl, rpm generated in Ubuntu
will mostly not work with Fedora/OpenSUSE.

Signed-off-by: Jorge Marques <jorge.marques@analog.com>