Remove workaround when checking for a trigger of a device
The bug where libiio would return error EIO in case the iio device
supports a trigger but none is assigned, no longer exists in latest
versions of libiio. Probably was fixed long time ago.
However I found that on serial backend the EIO is being returned by
iio_device_get_trigger() even though the intention was to report that
iio device doesn't support/need a trigger. The expected behaviour is
that 0 is returned and trigger argument is set to NULL. But this unhappy
coincidence was forcing users to provide a trigger for an iio device
that does not need one. It happened that no trigger was available making
the users unable to ask for samples from the device.
Note: this doesn't happend to all scenarios where serial backend is
used. I have seen the issue only on 2 projects. And I have seen it
working properly on other projects.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>