plugins: adrv9002: drop all the carrier save logic
Making sure that changing the carrier is properly handled is now done in
the kernel. Hence, all the logic in the plugin can be dropped.

The only thing the plugin now has to care about is to properly update
the value of the LO's that might automatically change value.

Signed-off-by: Nuno Sa <nuno.sa@analog.com>
(cherry picked from commit 6639701cf348fd25a2f2a3bfc41d6d8bb4363bbc)