plugins: adrv9002: fix out of bounds access
The array in 'struct carrier_helper' needs to be increased to the max
number of channels. The assumption of 3 being the max possible number
is no longer true. After
commit a2c666aa9ac58 ("plugins: adrv9002: fix carrier handling"), we can
have all ports moved to calibrated state using that helper struct (when all
ports are on the same LO). Hence, we could run into an out of bounds
access. On arch (as all packages and hence gtk are build with
-fstack-clash-protection) we, at least, get the program aborted with stack
smashing protection abort.

Signed-off-by: Nuno Sa <nuno.sa@analog.com>