iio_utils: add helper API to see if attribute is not present
Often, one wants to enable/disable certain features in a plugin
depending on the present or not of some attribute. hence, let's add an
helper for it. For now, this only supports checking device and channel
attributes as that should cover most of the cases.

Signed-off-by: Nuno Sa <nuno.sa@analog.com>
(cherry picked from commit 861b42e399d6bd046f0481ce4602bfeb41e4183c)