glade/oscplot: Fix runtime GTK3 warning when creating plot
The warning:
    (osc:17335): Gtk-WARNING **: 13:53:21.583: Content added to the action area of a dialog using header bars

The GtkDialogFileChooser that is being created via Glade has an built in GtkHeaderBar which has an action area where the 'Save' and 'Cancel' buttons are added. And this is not allowed, hence the above warning. The GTK2 version of the GtkDialogFileChooser didn't have a built in GtkHeaderBar.

The solution was to create in Glade a simple GtkDialog to which a GtkFileChooserWidget has been added. A GtkDialog does not have a built in GtkHeaderBar.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>