Revert "filters: adrv9002: add profiles for sdk 27.1"
This reverts commit 9b4291f45ea52c04cebc569704135b60c27439dd.

As it turns out, API 68.14.13 has no meaningful fixes and this release only
affects users using hdl that we do not use in the linux driver. Hence this
would only raise unnecessary questions without any added value and so it
was reverted from the driver. Therefore, let's revert it here as well.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>