oscmain: drop '-c' option
There's little advantage in having the '-c' option when compared with
'-u'. But more importantly, it was not working and leading to a segfault
since we we're not properly checking the return value. In libiio v1, we
get an encoded error code instead of NULL.

We could turn -c into a -u and pass that to iio_create_context() but it
is not worth the effort.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>