osc: fix crash when starting a plot capture
When starting a capture we we're getting a segfault because we we're
iterating over all IIO devices and trying to access things that could
not be available. Hence, make sure we only iterate over devices that are
effectively capable of buffering. For that, a new local helper function
was added. Obviously, this also has the advantage of not discarding
devices that do not matter making things a bit quicker.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>