dialogs: remove deprecated gdk_thread annotations
The version_info_show() function run on behalf of g_idle_add() which,
AFAIK, runs on the context of the main gtk thread. Therefore the
annotations were not needed anyways.

The scan context was also being done in a thread with no apparent
benefit. Hence, let's just run it from the main thread and remove the
deprecated annotations.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>