dialogs: improve app launch
Most of the times, we just start the application with the '-u' parameter
to connect to a specific device. Even if not, by default we start the
connection dialog with 'Manual' toggled. Therefore, there's no point in
doing an IIO scan blindly on app launch. That can add a very significant
overhead.i

Also instead of binding the GTK object to connect_clear(), bind it
directly to refresh_scan_ctx() so that we do a scan (which in turn
already calls connect_clear()).

Signed-off-by: Nuno Sá <nuno.sa@analog.com>