oscplot: fix use after free of 'priv->databox'
When a plot capture is "live", closing the windows (which destroys the
plot) leads to a use after in priv->databox because by the time our
redraw timer runs, that object is already freed. This does not
necessarily breaks at the first time, but try it enough times and
eventually osc gets a segfault.

Fix it by first checking the 'stop_redraw' flag.

As a bonus, there's also white space change/fix in this patch.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>