treewide: Move core osc to libiio v1
This moves all the core OSC source files to libiio v1.
Note that some utility function had to be added in process (to iio_utils.c).

For the plugins, the plan is to disable all of them (which this commit
is doing) and to convert one at a time in follow up patches.

Co-developed-by: Dan Nechita <dan.nechita@analog.com>
Signed-off-by: Dan Nechita <dan.nechita@analog.com>
Co-developed-by: Cristina Suteu <cristina.suteu@analog.com>
Signed-off-by: Cristina Suteu <cristina.suteu@analog.com>
Co-developed-by: Nuno Sá <nuno.sa@analog.com>
Signed-off-by: Nuno Sá <nuno.sa@analog.com>